backend/requirements.txt
annotated-types==0.7.0
appwrite==18.0.0
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.4.22
chardet==3.0.4
charset-normalizer==3.4.7
click==8.3.1
colorama==0.4.6
deep-translator==1.11.4
Deprecated==1.3.1
Flask==3.1.3
flask-cors==6.0.2
Flask-Limiter==4.1.1
googletrans==4.0.0rc1
gTTS==2.5.4
gunicorn==25.3.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
langdetect==1.0.9
legacy-cgi==2.6.4
limits==5.8.0
MarkupSafe==3.0.3
nltk==3.9.4
numpy==2.4.4
ordered-set==4.1.0
packaging==26.1
pydantic==2.13.3
pydantic_core==2.46.3
python-dotenv==1.2.2
regex==2026.4.4
requests==2.32.5
rfc3986==1.5.0
scikit-learn==1.8.0
scipy==1.17.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
threadpoolctl==3.6.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.8
Wikipedia-API==0.10.2
wrapt==2.1.2
